% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRBayes-package.R
\docType{package}
\name{DRBayes-package}
\alias{DRBayes}
\alias{DRBayes-package}
\title{DRBayes: Bayesian Doubly Robust Causal Inference Methods}
\description{
Bayesian doubly robust estimation of average treatment effects via posterior coupling, in which the outcome-model and propensity-score posteriors are linked so that the doubly robust moment condition is satisfied. Gibbs samplers for linear, logistic and probit models under conjugate or horseshoe priors are provided, and posterior draws produced by external Bayesian software can be supplied instead. A Bayesian bootstrap estimator and utilities for replicated simulation studies are also included. The methodology is described in Orihara, Momozaki and Sugasawa (2025) \doi{10.48550/arXiv.2506.04868}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://t-momozaki.github.io/DRBayes/}
  \item \url{https://github.com/t-momozaki/DRBayes}
  \item Report bugs at \url{https://github.com/t-momozaki/DRBayes/issues}
}

}
\author{
\strong{Maintainer}: Tomotaka Momozaki \email{t_momozaki@rs.tus.ac.jp}

Authors:
\itemize{
  \item Tomotaka Momozaki \email{t_momozaki@rs.tus.ac.jp}
  \item Shunichiro Orihara
  \item Shonosuke Sugasawa
}

}
\keyword{internal}
